# frontend/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OverlayPayload": {
        "description": "Predefined overlay entry returned by ``GET /overlays``.",
        "properties": {
          "name": {
            "description": "Display name of the overlay",
            "title": "Name",
            "type": "string"
          },
          "oid": {
            "description": "Overlay identifier",
            "title": "Oid",
            "type": "string"
          }
        },
        "required": [
          "name",
          "oid"
        ],
        "title": "OverlayPayload",
        "type": "object"
      },
      "OverlayStateUpdate": {
        "additionalProperties": true,
        "properties": {
          "match_info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchInfoModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "overlay_control": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OverlayControlModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "raw_remote_customization": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Raw Remote Customization"
          },
          "raw_remote_model": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Raw Remote Model"
          },
          "team_away": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TeamStateModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "team_home": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TeamStateModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "OverlayStateUpdate",
        "type": "object"
      },
      "RawConfigPayload": {
        "additionalProperties": true,
        "properties": {
          "customization": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Customization"
          },
          "model": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "title": "RawConfigPayload",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Volley Overlay Control",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/config/{overlay_id}": {
      "get": {
        "operationId": "get_config_api_config__overlay_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Config",
        "tags": [
          "Overlay"
        ]
      }
    },
    "/api/raw_config/{overlay_id}": {
      "get": {
        "operationId": "get_raw_config_api_raw_config__overlay_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Raw Config",
        "tags": [
          "Overlay"
        ]
      },
      "post": {
        "operationId": "set_raw_config_api_raw_config__overlay_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RawConfigPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Raw Config",
        "tags": [
          "Overlay"
        ]
      }
    },
    "/api/state/{overlay_id}": {
      "post": {
        "operationId": "update_state_api_state__overlay_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverlayStateUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update State",
        "tags": [
          "Overlay"
        ]
      }
    },
    "/api/theme/{overlay_id}/{theme_name}": {
      "post": {
        "operationId": "apply_theme_api_theme__overlay_id___theme_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "theme_name",
            "required": true,
            "schema": {
              "title": "Theme Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Theme",
        "tags": [
          "Overlay"
        ]
      }
    },
...
    "/api/v1/overlays": {
      "get": {
        "description": "Return predefined overlays available for selection.\n\nSourced exclusively from the ``PREDEFINED_OVERLAYS`` environment\nvariable (also populated via the remote configurator). Entries are\nfiltered by ``allowed_users`` using the caller's identity when user\nauthentication is enabled.",
        "operationId": "get_overlays_api_v1_overlays_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OverlayPayload"
                  },
                  "title": "Response Get Overlays Api V1 Overlays Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Overlays",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
...
    "/api/v1/themes": {
      "get": {
        "description": "Return available theme definitions.",
        "operationId": "get_themes_api_v1_themes_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Themes",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
    "/create/overlay/{overlay_id}": {
      "get": {
        "operationId": "create_overlay_create_overlay__overlay_id__get_post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Overlay",
        "tags": [
          "Overlay"
        ]
      },
      "post": {
        "operationId": "create_overlay_create_overlay__overlay_id__get_post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Overlay",
        "tags": [
          "Overlay"
        ]
      }
    },
    "/delete/overlay/{overlay_id}": {
      "delete": {
        "operationId": "delete_overlay_delete_overlay__overlay_id__delete_get_post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Overlay",
        "tags": [
          "Overlay"
        ]
      },
      "get": {
        "operationId": "delete_overlay_delete_overlay__overlay_id__delete_get_post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Overlay",
        "tags": [
          "Overlay"
        ]
      },
      "post": {
        "operationId": "delete_overlay_delete_overlay__overlay_id__delete_get_post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Overlay",
        "tags": [
          "Overlay"
        ]
      }
    },
...
    "/list/overlay": {
      "get": {
        "description": "Return every overlay id plus its output key.\n\nGated behind ``OVERLAY_MANAGER_PASSWORD`` because the response\ndefeats the capability-URL design of ``/overlay/{output_key}``.\nSee ``AUTHENTICATION.md`` (F-4).",
        "operationId": "list_overlays_list_overlay_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Overlays",
        "tags": [
          "Overlay"
        ]
      }
    },
...
    "/overlay/{overlay_id}": {
      "get": {
        "operationId": "serve_overlay_overlay__overlay_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "style",
            "required": false,
            "schema": {
              "title": "Style",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Serve Overlay",
        "tags": [
          "Overlay"
        ]
      }
    },
...
  }
}